requirements.txt
Pillow==5.4.1
django==2.1.5
django-ckeditor==5.6.1
django-analytical==2.5.0
django-maintenance-mode==0.13.0
django-maintenancemode==0.11.2
django-recaptcha==2.0.2
pylint==2.2.2